♻️ refactor util functions and constants
